Solution to problem 1

+++ 1.c
@@ -1,17 +1,16 @@
 #include <stdlib.h>
 #include <stdio.h>
-#include <inttypes.h>
 
 int main() {
-    uint64_t sum = 0;
+    int sum = 0;
 
-    for (uint16_t n = 0; n < 1000; n++) {
+    for (int n = 0; n < 1000; n++) {
         if (n % 3 == 0 || n % 5 == 0) {
             sum += n;
         }
     }
 
-    printf("%" PRIu64 "\n", sum);
+    printf("%d\n", sum);
 
     return EXIT_SUCCESS;
 }
